Implemented re-streaming of remote files
Implemented a basic system where files from remote hosts are downloaded in parts, then re-sent to the client using the Flask Responses and generators.

Currently, downloading a file via this re-streaming is iffy, since the client cannot do things like request certain sections. This means video content is almost entirely off the table (which is too bad).

In the future, I plan to implement some kind of cache system so the first user to download a remote file will have it streamed to them, but after that it will be served from a cache.This will result in less demand on the original remote, and let the user have better control of the process.